test: added test for post and reduce test: split deal ticket tests
